Fix HTTP body leak in fetchToken and preflight
An OK HTTP status would cause Body to be leaked (never closed). This
was encouraged by ResponseAsError, which required the user to
conditionally close the Body. Remove that footgun by unconditionally
closing the Body. This also happens to work better with GoLand's
analysis, gets rid of a warning, but it didn't catch these bugs.