cache: Reimplement files for insertion speed
Instead of a slice that's split with every op, this will use a linked
list along with a map to immediately look up any file by its path.

This improves performance from 1300ns/op to 50ns/op with a cache size
of 1000 entries. The old implementation performed linearly worse with
growing cache size, the new one becomes just marginally slower as it
grows, depending on Go map implementation details.

This loses the feature to insert with an arbitrary timestamp, but the
implementation did not really need that anyway.